Narrow nmCmdList 32-bit return signatures

+++ src/Event/mt_evtMain.c
@@ -7012,7 +7012,7 @@ void FUN_00366ac0(u32 param_1,u64 param_2)
 // FUN_00366EB0 NONMATCHING
 
 
-u64 FUN_00366eb0(u64 param_1,u64 param_2,u64 param_3)
+int FUN_00366eb0(int param_1,int param_2,int param_3)
